requirements.txt
# ============================================================================
# Core ML & Deep Learning
# ============================================================================
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ============================================================================
# Image Processing
# ============================================================================
pillow>=10.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# ============================================================================
# Web Framework - FastAPI (Production Ready)
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0

# ============================================================================
# WSGI Server for Production
# ============================================================================
gunicorn>=21.2.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
alembic>=1.12.0         # Database migrations

# ============================================================================
# Authentication & Security
# ============================================================================
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.1.0

# ============================================================================
# HTTP & API
# ============================================================================
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.25.0
factory-boy>=3.3.0

# ============================================================================
# Data Validation & Serialization
# ============================================================================
marshmallow>=3.20.0
jsonschema>=4.19.0

# ============================================================================
# Async Support
# ============================================================================
aiofiles>=23.2.0

# ============================================================================
# Logging & Monitoring
# ============================================================================
python-json-logger>=2.0.7
sentry-sdk>=1.38.0

# ============================================================================
# Date & Time
# ============================================================================
python-dateutil>=2.8.2
pytz>=2023.3

# ============================================================================
# Development & Code Quality
# ============================================================================
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
mypy>=1.7.0
pylint>=3.0.0

# ============================================================================
# Documentation
# ============================================================================
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# ============================================================================
# Optional but Recommended for Production
# ============================================================================
# AWS S3 Support
boto3>=1.28.0

# Cache Support
redis>=5.0.0

# Celery for async tasks
celery>=5.3.0
flake8>=6.1.0
mypy>=1.7.0
autopep8>=2.0.4
